Revert "Restructure and distribute functions #16"
This reverts commit afe0864ac7860a8b5cc8341c80dbe3f192e11cab.
